fixed pedantic warnings and some logic failures (e.g. check for >= 0 on usigned type)

+++ trunk/src/common/celgraph/celgraph.h
@@ -104,7 +104,7 @@ celPath, csObject, iCelPath, iComponent>
 {
  private:
   iObjectRegistry* object_reg;
-  size_t cur_node;
+  int cur_node;
   csRefArray<iMapNode> nodes;
   
  public:
